refactor(DUP): move `handle_internal` to its own module
Moves `handle_internal` form `Impl` to its own module to test it separately
