+++ apps/astarte_data_updater_plant/lib/astarte_data_updater_plant/data_updater/impl.ex
@@ -163,48 +163,8 @@ defmodule Astarte.DataUpdaterPlant.DataUpdater.Impl do
     %{new_state | connected: true, last_seen_message: timestamp}
   end
 
-  def handle_internal(state, "/heartbeat", _payload, message_id, timestamp) do
-    {:continue, handle_heartbeat(state, message_id, timestamp)}
-  end
-
-  def handle_internal(%State{discard_messages: true} = state, "/f", _, message_id, _) do
-    :ok = Queries.ack_end_device_deletion(state.realm, state.device_id)
-    _ = Logger.info("End device deletion acked.", tag: "device_delete_ack")
-    MessageTracker.ack_delivery(state.message_tracker, message_id)
-    {:stop, state}
-  end
-
   def handle_internal(state, path, payload, message_id, timestamp) do
-    Logger.warning(
-      "Unexpected internal message on #{path}, base64-encoded payload: #{inspect(Base.encode64(payload))}",
-      tag: "unexpected_internal_message"
-    )
-
-    {:ok, new_state} = Core.Device.ask_clean_session(state, timestamp)
-    MessageTracker.discard(new_state.message_tracker, message_id)
-
-    :telemetry.execute(
-      [:astarte, :data_updater_plant, :data_updater, :discarded_internal_message],
-      %{},
-      %{realm: new_state.realm}
-    )
-
-    base64_payload = Base.encode64(payload)
-
-    error_metadata = %{
-      "path" => inspect(path),
-      "base64_payload" => base64_payload
-    }
-
-    # TODO maybe we don't want triggers on unexpected internal messages?
-    Core.Trigger.execute_device_error_triggers(
-      new_state,
-      "unexpected_internal_message",
-      error_metadata,
-      timestamp
-    )
-
-    {:continue, Core.DataHandler.update_stats(new_state, "", nil, path, payload)}
+    Core.InternalHandler.handle_internal(state, path, payload, message_id, timestamp)
   end
 
   def start_device_deletion(state, timestamp) do
